Add trollsift to required packages and fix utils
