+++ nwcsafpps_runner/utils.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2018 - 2019 PyTroll
+# Copyright (c) 2018 - 2020 PyTroll
 
 # Author(s):
 
@@ -28,13 +28,17 @@ import stat
 import netifaces
 import shlex
 from glob import glob
+from subprocess import Popen, PIPE
 from posttroll.message import Message
 from trollsift.parser import parse
 import socket
 from six.moves.urllib.parse import urlparse
 from datetime import datetime, timedelta
 from nwcsafpps_runner.config import (LVL1_NPP_PATH, LVL1_EOS_PATH)
+import shlex
+import threading
 import logging
+
 LOG = logging.getLogger(__name__)
 
 
@@ -94,6 +98,29 @@ METOP_SENSOR = {'amsu-a': 'amsua', 'avhrr/3': 'avhrr',
 # METOP_NUMBER = {'b': '01', 'a': '02'}
 
 
+def run_command(cmdstr):
+    """Run system command"""
+    myargs = shlex.split(str(cmdstr))
+
+    LOG.debug("Command: " + str(cmdstr))
+    LOG.debug('Command sequence= ' + str(myargs))
+    try:
+        proc = Popen(myargs, shell=False, stderr=PIPE, stdout=PIPE)
+    except NwpPrepareError:
+        LOG.exception("Failed when preparing NWP data for PPS...")
+
+    out_reader = threading.Thread(
+        target=logreader, args=(proc.stdout, LOG.info))
+    err_reader = threading.Thread(
+        target=logreader, args=(proc.stderr, LOG.info))
+    out_reader.start()
+    err_reader.start()
+    out_reader.join()
+    err_reader.join()
+
+    return proc.wait()
+
+
 def check_uri(uri):
     """Check that the provided *uri* is on the local host and return the
     file path.
@@ -139,8 +166,8 @@ class SceneId(object):
 
     def __hash__(self):
         return hash(str(self.platform_name) + '_' +
-                str(self.orbit_number) + '_' +
-                str(self.starttime.strftime('%Y%m%d%H%M')))
+                    str(self.orbit_number) + '_' +
+                    str(self.starttime.strftime('%Y%m%d%H%M')))
 
     def __eq__(self, other):
 
@@ -316,7 +343,7 @@ def ready2run(msg, files4pps, **kwargs):
 
     LOG.debug("files4pps: %s", str(files4pps[sceneid]))
     if (stream_tag_name in msg.data and msg.data[stream_tag_name] in [stream_name, ] and
-        platform_name in SUPPORTED_METOP_SATELLITES):
+            platform_name in SUPPORTED_METOP_SATELLITES):
         LOG.info("EARS Metop data. Only require the HRPT/AVHRR level-1b file to be ready!")
     elif (platform_name in SUPPORTED_METOP_SATELLITES or
           platform_name in SUPPORTED_NOAA_SATELLITES):
